Domains Scraper 1.03

+++ Domains Scraper/Models/BacklinkType.cs
@@ -9,9 +9,9 @@ namespace Domains_Scraper.Models
     public class BacklinkType
     {
         public int Id { get; set; }
-        public int TextLinks { get; set; }
-        public int FrameLinks { get; set; }
-        public int FormLinks { get; set; }
-        public int ImageLinks { get; set; }
+        public long TextLinks { get; set; }
+        public long FrameLinks { get; set; }
+        public long FormLinks { get; set; }
+        public long ImageLinks { get; set; }
     }
 }
